refactor(inspector): FSD split for title bar and tabs visibility
- Add features/tabs-visibility (getTabsVisibility, useTabsVisibility, toggle UI)
- Add widgets/title-bar (TitleBar using tabs visibility)
- Simplify __root.tsx to layout, server start, SSE, settings modal only
- Import getTabsVisibility from feature in devtools page and routes
- Add use-tabs-visibility unit tests
