+++ packages/inspector/src/features/tabs-visibility/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Tabs visibility feature / 탭 표시 여부 기능
+ */
+export { getTabsVisibility, useTabsVisibility } from './model/use-tabs-visibility';
+export { TabVisibilityToggle } from './ui/tab-visibility-toggle';
