+++ packages/inspector/src/widgets/title-bar/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Title bar widget / 타이틀바 위젯
+ */
+export { TitleBar } from './ui/TitleBar';
